Successfully extract character creation and stat updates to a module
