+++ src/style.css
@@ -427,6 +427,7 @@ main {
 }
 
 td {
+  padding: 0.4rem;
   text-align: center;
 }
